requirements.txt
absl-py==2.0.0
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
ansi2html==1.8.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.0.0
cachetools==5.3.1
ccimport==0.4.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
crcmod==1.7
cryptography==41.0.4
cumm-cu116==0.4.11
cycler==0.11.0
Cython==3.0.2
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
-e git+https://github.com/MediaBrain-SJTU/CoCa3D@b39e5846bcd32dcf7cfed600fa126cff81f6dd36#egg=deformable_aggregation_ext&subdirectory=opencood/models/ops
defusedxml==0.7.1
descartes==1.1.0
distro==1.9.0
efficientnet-pytorch==0.7.1
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
fire==0.5.0
flake8==6.1.0
Flask==2.2.5
fonttools==4.42.1
fqdn==1.5.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
g2o-python==0.0.12
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.4
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.21.3
icecream==2.1.3
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lark==1.1.7
lazy_loader==0.3
llvmlite==0.41.0
lyft-dataset-sdk==0.0.8
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mat4py==0.6.0
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
mmcv==2.0.0rc4
mmdet==3.1.0
mmdet3d==1.2.0
mmengine==0.8.4
model-index==0.1.11
modelindex==0.0.2
mpi4py @ file:///croot/mpi4py_1671223370575/work
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.5.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.0
numpy==1.24.4
nuscenes-devkit==1.1.10
oauthlib==3.2.2
open3d==0.16.0
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.25
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pccm==0.4.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.3.0
plyfile==1.0.1
portalocker==2.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pypcd==0.1.1
pyquaternion==0.9.9
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lzf==0.2.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.2
scikit-build==0.17.6
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
Shapely==1.8.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spconv-cu116==2.3.6
sptr==0.0.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.16
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.3.3
tqdm==4.65.2
traitlets==5.10.0
trimesh==3.23.5
ttach==0.0.3
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.3
widgetsnbextension==4.0.9
yacs==0.1.8
yapf==0.40.1
zipp==3.16.2
